refactor: 컴포넌트명 변경
정해진 규칙대로 수정
